Split codegen expression literal tests
